extract iframe_test.go
